Blinking External LED Using External Button on STM32
This project demonstrates how to control an external LED using an external push button switch with an STM32 microcontroller. The LED will turn ON when the button is pressed and will turn OFF when the button is released. This simple application serves as a practical example of working with GPIO (General Purpose Input/Output) pins, debouncing techniques, and basic microcontroller programming.